Added --change-password functionality
When run, it will prompt for 2 passwords: the first decrypts the
keyboard input, and then reencrypts it with the new password.
The utility requires a lock on the database, so it requires selfspy
to not be running while it changes the passwords. Also, it does not
run the logger after automatically, because being explicit is
better.