Fixed sniff_x for the new activity_store
Moved a couple of things from the old activity_store into X sniffer.
Added a new hook to comply with new X sniffer.
Changed key_event to return with upper case.
No longer check for key up events as they are not logged.
button_event now returns the x and y coord for the press as per new
activity store.

Removed check for press on mouse click in activity_store as the
sniffers should not send it anyway.